backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pymongo>=4.6.0
yfinance>=0.2.0
requests>=2.31.0
feedparser>=6.0.0
apscheduler>=3.9.0
google-genai>=0.2.0
openai>=1.0.0
